Links go to separate pages now in navigation bar. Banner is moved to Home page. Hamburger menu appears normally and disappears when link is clicked. Collections is now own page.
