+++ apperals-frontend/src/components/Navigation.css
@@ -91,6 +91,7 @@ nav a:hover {
         color: white;
         /* padding: 10px 0; */
         text-align: left;
+        z-index: 1000;
     }
 
     .hamburger-active li {
