+++ apperals-frontend/src/components/pages/Collections.css
@@ -1,0 +1,20 @@
+.collections-container {
+    display: grid;
+    background-color: rgb(253, 102, 102);
+    grid-template-columns: repeat(3, 1fr);
+    gap: 16px;
+    padding: 16px;
+    width: 100%;
+}
+
+@media (max-width: 768px) {
+    .collections-container {
+        grid-template-columns: repeat(2, 1fr); /* 2 columns on medium screens */
+    }
+}
+
+@media (max-width: 480px) {
+    .collections-container {
+        grid-template-columns: 1fr; /* 1 column on small screens (mobile) */
+    }
+}
