+++ apperals-frontend/src/components/pages/Home.js
@@ -1,0 +1,13 @@
+import React from "react";
+import BGImage from "../BGImage";
+
+function Home() {
+    return (
+        <>
+            {/* Banner image */}
+            <BGImage />
+        </>
+    );
+}
+
+export default Home;
